requirements-supabase.txt
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
beautifulsoup4>=4.13.3
lxml>=5.3.1